requirements.txt
Flask==3.0.3
requests==2.32.3
twilio==9.3.0
APScheduler==3.10.4
python-dotenv==1.0.1
gunicorn==23.0.0
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9